Export data models at the root package level
Move the export of the classes and interfaces used as data models by the
components to the root level, since component classes can not be accessed
directly. This seems to be the convention adhered to by other packages.

Add a section to the README about consuming locally built distributions
of the library in other projects during development.